+++ rust/net/src/chat/error.rs
@@ -57,7 +57,10 @@ impl<T: Into<ConnectError>> From<TimeoutOr<RouteConnectError<T>>> for ConnectErr
             TimeoutOr::Other(RouteConnectError::AllAttemptsFailed) => {
                 ConnectError::AllAttemptsFailed
             }
-            TimeoutOr::Other(RouteConnectError::FatalConnect(err)) => err.into(),
+            TimeoutOr::Other(RouteConnectError::FatalConnect {
+                error,
+                failure_for_all_routes: _,
+            }) => error.into(),
             TimeoutOr::Timeout {
                 attempt_duration: _,
             } => ConnectError::Timeout,