+++ rust/net/src/connect_state.rs
@@ -339,16 +339,32 @@ impl<TC> ConnectionResources<'_, TC> {
                     Instant::now(),
                 );
                 log::debug!("[{log_tag}] connection attempt failed with {error}");
-                let is_fatal = match &mut error {
+                match &mut error {
                     WebSocketServiceConnectError::RejectedByServer {
                         response,
                         received_at: _,
                     } => {
                         log::trace!("[{log_tag}] full response: {response:?}");
+                        // 508 (with our confirmation header) is the server telling us to back off.
+                        if response.status().as_u16() == 508 {
+                            return ErrorHandling::Fatal {
+                                error,
+                                failure_for_all_routes: Some(UnsuccessfulOutcome::ShortTerm),
+                            };
+                        }
+
                         // Retry-After takes precedence over everything else.
-                        libsignal_net_infra::extract_retry_later(response.headers()).is_some() ||
-                        // If we're rejected based on the request (4xx), there's no point in retrying.
-                        response.status().is_client_error()
+                        // And if we're rejected based on the request (4xx), there's no point in
+                        // continuing (though this shouldn't really happen for most of our
+                        // services).
+                        if libsignal_net_infra::extract_retry_later(response.headers()).is_some()
+                            || response.status().is_client_error()
+                        {
+                            return ErrorHandling::Fatal {
+                                error,
+                                failure_for_all_routes: None,
+                            };
+                        }
                     }
                     WebSocketServiceConnectError::Connect(
                         connect_error,
@@ -359,7 +375,12 @@ impl<TC> ConnectionResources<'_, TC> {
                             // as fatal.
                             WebSocketConnectError::Transport(
                                 TransportConnectError::ClientAbort,
-                            ) => true,
+                            ) => {
+                                return ErrorHandling::Fatal {
+                                    error,
+                                    failure_for_all_routes: None,
+                                };
+                            }
                             // An unexpected self-signed certificate suggests a traffic-inspecting
                             // proxy of some kind. This isn't a *fatal* error; we still want to try
                             // other routes. However, if no routes succeed nor have a more specific
@@ -379,15 +400,11 @@ impl<TC> ConnectionResources<'_, TC> {
                             }
                             // In any other case, if we didn't make it to the server, we should
                             // retry.
-                            _ => false,
+                            _ => {}
                         }
                     }
-                };
-                if is_fatal {
-                    ErrorHandling::Fatal(error)
-                } else {
-                    ErrorHandling::Continue
                 }
+                ErrorHandling::Continue
             },
         )
         .await
@@ -428,7 +445,10 @@ impl<TC> ConnectionResources<'_, TC> {
                 | TimeoutOr::Timeout {
                     attempt_duration: _,
                 } => crate::enclave::Error::AllConnectionAttemptsFailed,
-                TimeoutOr::Other(ConnectError::FatalConnect(e)) => e.into(),
+                TimeoutOr::Other(ConnectError::FatalConnect {
+                    error,
+                    failure_for_all_routes: _,
+                }) => error.into(),
             })?;
 
         let connection =
@@ -481,7 +501,10 @@ impl<TC> ConnectionResources<'_, TC> {
                 // any failures as route-specific (other than choosing to abort).
                 match error {
                     HttpConnectError::Transport(TransportConnectError::ClientAbort) => {
-                        ErrorHandling::Fatal(error)
+                        ErrorHandling::Fatal {
+                            error,
+                            failure_for_all_routes: None,
+                        }
                     }
                     HttpConnectError::Transport(TransportConnectError::SslFailedHandshake(
                         reason,
@@ -587,21 +610,56 @@ impl<TC> ConnectionResources<'_, TC> {
                 attempt_duration: connect_timeout,
             })?;
 
-        match &result {
-            Ok((_connection, route)) => log::info!(
-                "[{log_tag}] connection through {route} succeeded after {:.3?}",
-                updates.finished_at - start
-            ),
-            Err(e) => log::info!("[{log_tag}] connection failed with {e}"),
+        let service_level_outcome = match &result {
+            Ok((_connection, route)) => {
+                log::info!(
+                    "[{log_tag}] connection through {route} succeeded after {:.3?}",
+                    updates.finished_at - start
+                );
+                Some(AttemptOutcome {
+                    started: start,
+                    result: Ok(()),
+                })
+            }
+            Err(e) => {
+                log::info!("[{log_tag}] connection failed with {e}");
+                if let ConnectError::FatalConnect {
+                    error: _,
+                    failure_for_all_routes: Some(outcome),
+                } = e
+                {
+                    Some(AttemptOutcome {
+                        started: start,
+                        result: Err(*outcome),
+                    })
+                } else {
+                    None
+                }
+            }
+        };
+
+        let per_route_outcomes = process_outcomes(updates.outcomes);
+        let system_now = SystemTime::now();
+
+        {
+            let mut connect_state_guard = connect_state.lock().expect("not poisoned");
+            connect_state_guard.attempts_record.apply_outcome_updates(
+                per_route_outcomes,
+                updates.finished_at,
+                system_now,
+            );
+            if let Some(outcome) = service_level_outcome {
+                // Apply failures doubly so we back off more aggressively. (We could also do this by
+                // having two sets of ConnectionOutcomeParams, but that's harder to keep track of.)
+                connect_state_guard
+                    .service_level_attempts_record
+                    .apply_outcome_updates(
+                        [(service, outcome), (service, outcome)],
+                        updates.finished_at,
+                        system_now,
+                    );
+            }
         }
-
-        let outcomes = process_outcomes(updates.outcomes);
-
-        connect_state
-            .lock()
-            .expect("not poisoned")
-            .attempts_record
-            .apply_outcome_updates(outcomes, updates.finished_at, SystemTime::now());
 
         let (connection, description) = result?;
         Ok((
@@ -751,9 +809,10 @@ where
             log_tag,
             |error| {
                 match error {
-                    InterfaceChangedOr::InterfaceChanged => {
-                        ErrorHandling::Fatal(TransportConnectError::ClientAbort)
-                    }
+                    InterfaceChangedOr::InterfaceChanged => ErrorHandling::Fatal {
+                        error: TransportConnectError::ClientAbort,
+                        failure_for_all_routes: None,
+                    },
                     InterfaceChangedOr::Other(_) => {
                         // All normal transport-level errors are considered intermittent; see
                         // WebSocketServiceConnectError::classify.
@@ -861,6 +920,7 @@ mod test {
         TcpRoute, TlsRoute, TlsRouteFragment, UnresolvedHost, UnresolvedTransportRoute,
         UnsuccessfulOutcome, WebSocketRoute,
     };
+    use libsignal_net_infra::testutil::TestError;
     use libsignal_net_infra::utils::no_network_change_events;
     use libsignal_net_infra::{Alpn, OverrideNagleAlgorithm, RouteType};
     use nonzero_ext::nonzero;
@@ -1039,6 +1099,194 @@ mod test {
     }
 
     #[tokio::test(start_paused = true)]
+    async fn service_wide_error_is_recorded() {
+        let service = ServiceName("test");
+        let start = Instant::now();
+
+        let ws_connector = <crate::infra::ws::Stateless>::default();
+        let resolver = DnsResolver::new_from_static_map(HashMap::from([(
+            FAKE_HOST_NAME,
+            LookupResult::new(vec![ip_addr!(v4, "192.0.2.1")], vec![]),
+        )]));
+
+        let connector = ConnectFn(|(), _| {
+            std::future::ready(Err::<tokio::io::DuplexStream, _>(
+                TransportConnectError::TcpConnectionFailed,
+            ))
+        });
+
+        let state = ConnectState {
+            connect_timeout: Duration::MAX,
+            network_interface_poll_interval: Duration::MAX,
+            post_route_change_connect_timeout: Duration::MAX,
+            route_resolver: RouteResolver::default(),
+            attempts_record: ConnectionOutcomes::new(SUGGESTED_CONNECT_PARAMS),
+            service_level_attempts_record: ConnectionOutcomes::new(SUGGESTED_CONNECT_PARAMS),
+            make_transport_connector: connector,
+            route_provider_context: Default::default(),
+        };
+
+        assert_eq!(
+            state
+                .service_level_attempts_record
+                .compute_delay(&service, start),
+            Duration::ZERO
+        );
+
+        let state = state.into();
+
+        let connection_resources = ConnectionResources {
+            connect_state: &state,
+            dns_resolver: &resolver,
+            network_change_event: &no_network_change_events(),
+            confirmation_header_name: None,
+        };
+
+        let connect = connection_resources.connect_over_transport(
+            service,
+            vec![FAKE_WEBSOCKET_ROUTES[0].clone()],
+            ws_connector,
+            "test",
+            |_error| ErrorHandling::Fatal {
+                error: TestError::Expected,
+                failure_for_all_routes: Some(UnsuccessfulOutcome::ShortTerm),
+            },
+        );
+
+        let result: Result<_, TimeoutOr<ConnectError<_>>> = connect.await;
+
+        assert_matches!(
+            result,
+            Err(TimeoutOr::Other(ConnectError::FatalConnect {
+                error: TestError::Expected,
+                failure_for_all_routes: Some(UnsuccessfulOutcome::ShortTerm)
+            }))
+        );
+
+        let mut state = state.lock().expect("not poisoned");
+        assert_ne!(
+            state
+                .service_level_attempts_record
+                .compute_delay(&service, start),
+            Duration::ZERO
+        );
+        assert_eq!(
+            state.service_level_attempts_record.compute_delay(
+                &service,
+                start + SUGGESTED_CONNECT_PARAMS.short_term_age_cutoff
+            ),
+            Duration::ZERO
+        );
+
+        // Network changes don't reset service-level delays (since they're assumed to not be about
+        // the client side at all).
+        state.network_changed(start);
+        assert_ne!(
+            state
+                .service_level_attempts_record
+                .compute_delay(&service, start),
+            Duration::ZERO
+        );
+        assert_eq!(
+            state.service_level_attempts_record.compute_delay(
+                &service,
+                start + SUGGESTED_CONNECT_PARAMS.short_term_age_cutoff
+            ),
+            Duration::ZERO
+        );
+    }
+
+    #[tokio::test(start_paused = true)]
+    async fn ws_508_is_a_service_wide_error() {
+        let service = ServiceName("test");
+        let start = Instant::now();
+
+        let resolver = DnsResolver::new_from_static_map(HashMap::from([(
+            FAKE_HOST_NAME,
+            LookupResult::new(vec![ip_addr!(v4, "192.0.2.1")], vec![]),
+        )]));
+
+        let connector = ConnectFn(|(), _| std::future::ready(Ok::<_, WebSocketConnectError>(())));
+        let ws_connector = ConnectFn(|(), _| {
+            std::future::ready(Err::<tokio::io::DuplexStream, _>(
+                WebSocketConnectError::WebSocketError(
+                    libsignal_net_infra::ws::WebSocketError::Http(
+                        http::Response::builder()
+                            .status(508)
+                            .header("confirmation-header", "")
+                            .body(None)
+                            .expect("valid")
+                            .into(),
+                    ),
+                ),
+            ))
+        });
+
+        let state = ConnectState {
+            connect_timeout: Duration::MAX,
+            network_interface_poll_interval: Duration::MAX,
+            post_route_change_connect_timeout: Duration::MAX,
+            route_resolver: RouteResolver::default(),
+            attempts_record: ConnectionOutcomes::new(SUGGESTED_CONNECT_PARAMS),
+            service_level_attempts_record: ConnectionOutcomes::new(SUGGESTED_CONNECT_PARAMS),
+            make_transport_connector: connector,
+            route_provider_context: Default::default(),
+        };
+
+        assert_eq!(
+            state
+                .service_level_attempts_record
+                .compute_delay(&service, start),
+            Duration::ZERO
+        );
+
+        let state = state.into();
+
+        let connection_resources = ConnectionResources {
+            connect_state: &state,
+            dns_resolver: &resolver,
+            network_change_event: &no_network_change_events(),
+            confirmation_header_name: Some(http::HeaderName::from_static("confirmation-header")),
+        };
+
+        let connect = connection_resources.connect_ws(
+            service,
+            vec![FAKE_WEBSOCKET_ROUTES[0].clone()],
+            ws_connector,
+            "test",
+        );
+
+        let result: Result<_, TimeoutOr<ConnectError<_>>> = connect.await;
+
+        assert_matches!(
+            result,
+            Err(TimeoutOr::Other(ConnectError::FatalConnect {
+                error: WebSocketServiceConnectError::RejectedByServer {
+                    response,
+                    received_at,
+                },
+                failure_for_all_routes: Some(UnsuccessfulOutcome::ShortTerm)
+            }))
+            if received_at == start && response.status().as_u16() == 508
+        );
+
+        let state = state.lock().expect("not poisoned");
+        assert_ne!(
+            state
+                .service_level_attempts_record
+                .compute_delay(&service, start),
+            Duration::ZERO
+        );
+        assert_eq!(
+            state.service_level_attempts_record.compute_delay(
+                &service,
+                start + SUGGESTED_CONNECT_PARAMS.short_term_age_cutoff
+            ),
+            Duration::ZERO
+        );
+    }
+
+    #[tokio::test(start_paused = true)]
     async fn client_abort_transport_error_is_fatal() {
         // We can't directly test the ClientAbort produced for a network change without *more*
         // custom dependency injection for connect_ws---we can fire the network change event, but we
@@ -1091,12 +1339,13 @@ mod test {
 
         assert_matches!(
             result,
-            Err(TimeoutOr::Other(ConnectError::FatalConnect(
-                WebSocketServiceConnectError::Connect(
+            Err(TimeoutOr::Other(ConnectError::FatalConnect {
+                error: WebSocketServiceConnectError::Connect(
                     WebSocketConnectError::Transport(TransportConnectError::ClientAbort),
                     NotRejectedByServer { .. }
-                )
-            )))
+                ),
+                failure_for_all_routes: None,
+            }))
         );
     }
 
@@ -1162,8 +1411,8 @@ mod test {
 
         assert_matches!(
             result,
-            Err(TimeoutOr::Other(ConnectError::FatalConnect(
-                WebSocketServiceConnectError::Connect(
+            Err(TimeoutOr::Other(ConnectError::FatalConnect {
+                error: WebSocketServiceConnectError::Connect(
                     WebSocketConnectError::Transport(TransportConnectError::SslFailedHandshake(
                         FailedHandshakeReason::Cert {
                             error: X509VerifyError::SELF_SIGNED_CERT_IN_CHAIN,
@@ -1171,8 +1420,9 @@ mod test {
                         }
                     )),
                     NotRejectedByServer { .. }
-                )
-            )))
+                ),
+                failure_for_all_routes: None,
+            }))
         );
     }
 
@@ -1584,9 +1834,10 @@ mod test {
 
         assert_matches!(
             result,
-            Err(TimeoutOr::Other(ConnectError::FatalConnect(
-                HttpConnectError::Transport(TransportConnectError::ClientAbort),
-            )))
+            Err(TimeoutOr::Other(ConnectError::FatalConnect {
+                error: HttpConnectError::Transport(TransportConnectError::ClientAbort),
+                failure_for_all_routes: None,
+            }))
         );
     }